+++ packages/terra-list/CHANGELOG.md
@@ -1,6 +1,11 @@
 # Changelog
 
 ## Unreleased
+
+* Added
+  * Added arrow key navigation support for list items.
+## 4.63.0 - (September 26, 2023)
+
 * Added
   * Added drag and drop icon for draggable list items.